@@ -1,8 +1,12 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-	content: ["./src/**/*.{vue,html,js}"],
+// const defaultTheme = require("tailwindcss/defaultTheme")
+export default {
+	content: ["./index.html", "./src/**/*.{vue, js, jsx, html, ts, tsx"],
 	theme: {
 		extend: {
+			// fontFamily: {
+			// 	sans: ['InterVariable', ...defaultTheme.fontFamily.sans]
+			// }
 		},
 	},
 	plugins: [],
