@@ -1,31 +1,19 @@
-import { createRouter, createWebHistory } from 'vue-router'
-import Home from '../views/Home.vue'
-import ScgaList from '@/components/ScgaList.vue'
+import { createRouter, createWebHistory } from "vue-router";
 
 const routes = [
-  {
-    path: '/',
-    name: 'home',
-    component: Home
-  },
-  {
-    path: '/scgas',
-    name: 'SCGA',
-    component: ScgaList
-  },
-  {
-    path: '/about',
-    name: 'about',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-  }
-]
+	{
+		path: "/",
+		alias: "/scga",
+		// redirect: '/',
+		name: 'scga',
+		component: () => import ("@/views/scga.vue"),
+	},
+	
+];
 
 const router = createRouter({
-  history: createWebHistory(process.env.BASE_URL),
-  routes
-})
+	history: createWebHistory(),
+	routes,
+});
 
-export default router
+export default router;
